+++ src/forms/editEntityForm.tsx
@@ -1,283 +1,271 @@
 import classnames from 'classnames';
 import * as React from 'react';
 
+import { mapAbortedToNull } from '../coreUtils/async';
 import { useKeyedSyncStore } from '../coreUtils/keyedObserver';
+import { useTranslation } from '../coreUtils/i18n';
 
 import { ElementModel, ElementIri, PropertyTypeIri } from '../data/model';
 import * as Rdf from '../data/rdf/rdfModel';
+import type {
+    MetadataEntityShape, MetadataPropertyShape, MetadataValueShape,
+} from '../data/metadataProvider';
+
+import { HtmlSpinner } from '../diagram/spinner';
 
 import { subscribePropertyTypes } from '../editor/observedElement';
 import { WithFetchStatus } from '../editor/withFetchStatus';
 
-import { WorkspaceContext, useWorkspace } from '../workspace/workspaceContext';
+import { useWorkspace } from '../workspace/workspaceContext';
+
+import { PropertyUpdater, TextPropertyInput } from './textPropertyInput';
 
 const FORM_CLASS = 'reactodia-form';
 const CLASS_NAME = 'reactodia-edit-entity-form';
 
-export interface EditEntityFormProps {
+export function EditEntityForm(props: {
     entity: ElementModel;
     onApply: (entity: ElementModel) => void;
     onCancel: () => void;
-}
+}) {
+    const {entity, onApply, onCancel} = props;
+    const {model, editor, translation: t} = useWorkspace();
 
-interface State {
-    elementModel: ElementModel;
-}
+    const [data, setData] = React.useState(entity);
+    const [shape, setShape] = React.useState<MetadataEntityShape>();
+    const [error, setError] = React.useState<unknown>();
 
-export class EditEntityForm extends React.Component<EditEntityFormProps, State> {
-    static contextType = WorkspaceContext;
-    declare readonly context: WorkspaceContext;
-
-    constructor(props: EditEntityFormProps, context: any) {
-        super(props, context);
-        this.state = {elementModel: props.entity};
-    }
-
-    componentDidUpdate(prevProps: EditEntityFormProps) {
-        if (this.props.entity !== prevProps.entity) {
-            this.setState({elementModel: this.props.entity});
+    React.useEffect(() => {
+        setData(entity);
+        if (editor.metadataProvider) {
+            setShape(undefined);
+            setError(undefined);
+            const cancellation = new AbortController();
+            const signal = cancellation.signal;
+            mapAbortedToNull(
+                editor.metadataProvider.getEntityShape(entity.types, {signal}),
+                signal
+            ).then(
+                result => {
+                    if (result === null) {
+                        return;
+                    }
+                    setShape(result);
+                },
+                error => setError(error)
+            );
+        } else {
+            setShape(DEFAULT_ENTITY_SHAPE);
         }
-    }
+    }, [entity]);
 
-    private renderType() {
-        const {model, translation: t} = this.context;
-        const {elementModel} = this.state;
-        return (
-            <label>
-                {t.text('visual_authoring.edit_entity.type.label')}
-                {elementModel.types.map(type => (
-                    <input key={type}
-                        className='reactodia-form-control'
-                        name='reactodia-edit-entity-type'
-                        title={type}
-                        value={t.formatLabel(
-                            model.getElementType(type)?.data?.label, type, model.language
-                        )}
-                        disabled={true}
-                    />
-                ))}
-            </label>
-        );
-    }
+    const languages = React.useMemo(
+        () => editor.metadataProvider?.getLiteralLanguages() ?? [], []
+    );
 
-    private onChangeIri = (e: React.FormEvent<HTMLInputElement>) => {
+    const onChangeIri = React.useCallback((e: React.FormEvent<HTMLInputElement>) => {
         const target = (e.target as HTMLInputElement);
         const iri = target.value as ElementIri;
-        this.setState(prevState => {
+        setData(previous => ({...previous, id: iri}));
+    }, []);
+
+    const onChangeLabel = React.useCallback((updater: PropertyUpdater) => {
+        setData(previous => ({
+            ...previous,
+            label: updater(previous.label) as ReadonlyArray<Rdf.Literal>,
+        }));
+    }, []);
+
+    const onChangeProperty = (
+        property: PropertyTypeIri,
+        updater: PropertyUpdater
+    ): void => {
+        setData(previous => {
+            const properties = previous.properties;
+            const values = Object.prototype.hasOwnProperty.call(properties, property)
+                ? properties[property] : undefined;
             return {
-                elementModel: {
-                    ...prevState.elementModel,
-                    id: iri,
+                ...previous,
+                properties: {
+                    ...properties,
+                    [property]: updater(values ?? []),
                 }
             };
         });
     };
 
-    private renderIri() {
-        const {translation: t} = this.context;
-        const {elementModel} = this.state;
+    if (!shape) {
         return (
-            <label>
-                {t.text('visual_authoring.edit_entity.iri.label')}
-                <input className='reactodia-form-control'
-                    name='reactodia-edit-entity-iri'
-                    defaultValue={elementModel.id}
-                    onChange={this.onChangeIri}
+            <div className={classnames(FORM_CLASS, CLASS_NAME, `${CLASS_NAME}--loading`)}>
+                <HtmlSpinner width={30} height={30}
+                    errorOccurred={Boolean(error)}
                 />
-            </label>
-        );
-    }
-
-    private onChangeLabel = (e: React.FormEvent<HTMLInputElement>) => {
-        const {model} = this.context;
-        const target = (e.target as HTMLInputElement);
-        const labels = target.value.length > 0 ? [model.factory.literal(target.value)] : [];
-        this.setState({elementModel: {
-            ...this.state.elementModel,
-            label: labels,
-        }});
-    };
-
-    private renderLabel() {
-        const {model, translation: t} = this.context;
-        const label = t.selectLabel(this.state.elementModel.label, model.language);
-        const text = label ? label.value : '';
-        return (
-            <label>
-                {t.text('visual_authoring.edit_entity.label.label')}
-                <input className='reactodia-form-control'
-                    name='reactodia-edit-entity-label'
-                    value={text}
-                    onChange={this.onChangeLabel}
-                />
-            </label>
-        );
-    }
-
-    render() {
-        const {translation: t} = this.context;
-        const {elementModel} = this.state;
-        return (
-            <div className={classnames(FORM_CLASS, CLASS_NAME)}>
-                <div className={`reactodia-scrollable ${FORM_CLASS}__body`}>
-                    <div className={`${FORM_CLASS}__row`}>
-                        {this.renderIri()}
-                    </div>
-                    <div className={`${FORM_CLASS}__row`}>
-                        {this.renderType()}
-                    </div>
-                    <div className={`${FORM_CLASS}__row`}>
-                        {this.renderLabel()}
-                    </div>
-                    <Properties data={elementModel}
-                        onChangeProperty={this.onChangeProperty}
-                    />
-                </div>
-                <div className={`${FORM_CLASS}__controls`}>
-                    <button type='button'
-                        className={`reactodia-btn reactodia-btn-primary ${FORM_CLASS}__apply-button`}
-                        title={t.text('visual_authoring.dialog.apply.title')}
-                        onClick={() => this.props.onApply(this.state.elementModel)}>
-                        {t.text('visual_authoring.dialog.apply.label')}
-                    </button>
-                    <button type='button'
-                        className='reactodia-btn reactodia-btn-secondary'
-                        title={t.text('visual_authoring.dialog.cancel.title')}
-                        onClick={this.props.onCancel}>
-                        {t.text('visual_authoring.dialog.cancel.label')}
-                    </button>
-                </div>
             </div>
         );
     }
 
-    private onChangeProperty = (
-        property: PropertyTypeIri,
-        values: ReadonlyArray<Rdf.NamedNode | Rdf.Literal>
-    ): void => {
-        this.setState(prevState => {
-            return {
-                elementModel: {
-                    ...prevState.elementModel,
-                    properties: {
-                        ...prevState.elementModel.properties,
-                        [property]: values,
-                    }
-                }
-            };
-        });
-    };
+    return (
+        <div className={classnames(FORM_CLASS, CLASS_NAME)}>
+            <div className={`reactodia-scrollable ${FORM_CLASS}__body`}>
+                <div className={`${FORM_CLASS}__row`}>
+                    <label>
+                        {t.text('visual_authoring.edit_entity.iri.label')}
+                        <input className='reactodia-form-control'
+                            name='reactodia-edit-entity-iri'
+                            defaultValue={data.id}
+                            onChange={onChangeIri}
+                        />
+                    </label>
+                </div>
+                <div className={`${FORM_CLASS}__row`}>
+                    <label>
+                        {t.text('visual_authoring.edit_entity.type.label')}
+                        {data.types.map(type => (
+                            <input key={type}
+                                className='reactodia-form-control'
+                                name='reactodia-edit-entity-type'
+                                title={type}
+                                value={t.formatLabel(
+                                    model.getElementType(type)?.data?.label, type, model.language
+                                )}
+                                disabled={true}
+                            />
+                        ))}
+                    </label>
+                </div>
+                <div className={`${FORM_CLASS}__row`}>
+                    <label>{t.text('visual_authoring.edit_entity.label.label')}</label>
+                    <TextPropertyInput shape={DEFAULT_PROPERTY_SHAPE}
+                        languages={languages}
+                        values={data.label}
+                        updateValues={onChangeLabel}
+                        factory={model.factory}
+                    />
+                </div>
+                <Properties properties={shape.properties}
+                    languages={languages}
+                    data={data.properties}
+                    onChangeData={onChangeProperty}
+                />
+            </div>
+            <div className={`${FORM_CLASS}__controls`}>
+                <button type='button'
+                    className={`reactodia-btn reactodia-btn-primary ${FORM_CLASS}__apply-button`}
+                    title={t.text('visual_authoring.dialog.apply.title')}
+                    onClick={() => onApply(data)}>
+                    {t.text('visual_authoring.dialog.apply.label')}
+                </button>
+                <button type='button'
+                    className='reactodia-btn reactodia-btn-secondary'
+                    title={t.text('visual_authoring.dialog.cancel.title')}
+                    onClick={onCancel}>
+                    {t.text('visual_authoring.dialog.cancel.label')}
+                </button>
+            </div>
+        </div>
+    );
 }
 
+const DEFAULT_ENTITY_SHAPE: MetadataEntityShape = {
+    properties: new Map(),
+};
+const DEFAULT_VALUE_SHAPE: MetadataValueShape = {
+    termType: 'Literal',
+};
+const DEFAULT_PROPERTY_SHAPE: MetadataPropertyShape = {
+    valueShape: DEFAULT_VALUE_SHAPE,
+};
 
 function Properties(props: {
-    data: ElementModel;
-    onChangeProperty: (
-        property: PropertyTypeIri,
-        values: ReadonlyArray<Rdf.NamedNode | Rdf.Literal>
-    ) => void;
+    properties: ReadonlyMap<PropertyTypeIri, MetadataPropertyShape>;
+    languages: ReadonlyArray<string>;
+    data: { readonly [id: string]: ReadonlyArray<Rdf.NamedNode | Rdf.Literal> };
+    onChangeData: (property: PropertyTypeIri, updater: PropertyUpdater) => void;
 }) {
-    const {data, onChangeProperty} = props;
+    const {properties, languages, data, onChangeData} = props;
     const {model, translation: t} = useWorkspace();
 
-    const propertyIris = Object.keys(data.properties) as PropertyTypeIri[];
+    const extendedProperties = new Map(properties);
+    for (const propertyIri of Object.keys(data) as PropertyTypeIri[]) {
+        if (!extendedProperties.has(propertyIri)) {
+            extendedProperties.set(propertyIri, DEFAULT_PROPERTY_SHAPE);
+        }
+    }
+    const propertyIris = Array.from(extendedProperties.keys());
     useKeyedSyncStore(subscribePropertyTypes, propertyIris, model);
 
     if (propertyIris.length === 0) {
         return null;
     }
 
-    const properties = propertyIris.map(iri => {
+    const labelledProperties = Array.from(extendedProperties, ([iri, shape]) => {
         const property = model.getPropertyType(iri);
+        const values = Object.prototype.hasOwnProperty.call(data, iri) ? data[iri] : undefined;
         return {
             iri,
             label: t.formatLabel(property?.data?.label, iri, model.language),
-            values: data.properties[iri],
+            shape,
+            values: values ?? [],
         };
     });
-    properties.sort((a, b) => a.label.localeCompare(b.label));
+    labelledProperties.sort((a, b) => a.label.localeCompare(b.label));
 
     return (
         <div role='list'
             className={`${CLASS_NAME}__properties`}>
-            {properties.map(({iri, label, values}) => {
-                return (
-                    <div key={iri} className={`${FORM_CLASS}__row`}>
-                        <label
-                            title={t.text('visual_authoring.edit_entity.property.title', {
-                                property: label,
-                                propertyIri: iri,
-                            })}>
-                            <WithFetchStatus type='propertyType' target={iri}>
-                                <span>
-                                    {t.text('visual_authoring.edit_entity.property.label', {
-                                        property: label,
-                                        propertyIri: iri,
-                                    })}
-                                </span>
-                            </WithFetchStatus>
-                            <PropertyValues property={iri}
-                                values={values}
-                                onChange={onChangeProperty}
-                                factory={model.factory}
-                            />
-                        </label>
-                    </div>
-                );
-            })}
+            {labelledProperties.map(({iri, label, shape, values}) =>
+                <Property key={iri}
+                    iri={iri}
+                    label={label}
+                    shape={shape}
+                    languages={languages}
+                    values={values}
+                    onChange={onChangeData}
+                    factory={model.factory}
+                />
+            )}
         </div>
     );
 }
 
-function PropertyValuesInner(props: {
-    property: PropertyTypeIri;
+function Property(props: {
+    iri: PropertyTypeIri;
+    label: string;
+    shape: MetadataPropertyShape;
+    languages: ReadonlyArray<string>;
     values: ReadonlyArray<Rdf.NamedNode | Rdf.Literal>;
-    onChange: (
-        property: PropertyTypeIri,
-        values: ReadonlyArray<Rdf.NamedNode | Rdf.Literal>
-    ) => void;
+    onChange: (iri: PropertyTypeIri, updater: PropertyUpdater) => void;
     factory: Rdf.DataFactory;
 }) {
-    const {property, values, onChange, factory} = props;
+    const {iri, label, shape, languages, values, onChange, factory} = props;
+    const t = useTranslation();
+
+    const updateValues = React.useCallback((updater: PropertyUpdater) => {
+        onChange(iri, updater);
+    }, [iri, onChange]);
+
     return (
-        <>
-            {values.map((term, index) => (
-                <input key={index}
-                    name='reactodia-edit-entity-property'
-                    className={classnames('reactodia-form-control')}
-                    value={term.value}
-                    onChange={e => {
-                        const changedValue = e.currentTarget.value;
-                        const nextValues = [...values];
-                        nextValues[index] = setTermValue(term, changedValue, factory);
-                        onChange(property, nextValues);
-                    }}
-                />
-            ))}
-        </>
+        <div className={`${FORM_CLASS}__row`}>
+            <label
+                title={t.text('visual_authoring.property.title', {
+                    property: label,
+                    propertyIri: iri,
+                })}>
+                <WithFetchStatus type='propertyType' target={iri}>
+                    <span>
+                        {t.text('visual_authoring.property.label', {
+                            property: label,
+                            propertyIri: iri,
+                        })}
+                    </span>
+                </WithFetchStatus>
+            </label>
+            <TextPropertyInput shape={shape}
+                languages={languages}
+                values={values}
+                updateValues={updateValues}
+                factory={factory}
+            />
+        </div>
     );
 }
-
-function setTermValue(
-    term: Rdf.NamedNode | Rdf.Literal,
-    value: string,
-    factory: Rdf.DataFactory
-): Rdf.NamedNode | Rdf.Literal {
-    if (term.termType === 'NamedNode') {
-        return factory.namedNode(value);
-    } else if (term.language) {
-        return factory.literal(value, term.language);
-    } else {
-        return factory.literal(value, term.datatype);
-    }
-}
-
-const PropertyValues = React.memo(
-    PropertyValuesInner,
-    (prevProps, nextProps) => (
-        prevProps.property === nextProps.property &&
-        prevProps.values === nextProps.values &&
-        prevProps.onChange === nextProps.onChange &&
-        prevProps.factory === nextProps.factory
-    )
-);