+++ src/workspace.ts
@@ -22,7 +22,7 @@ export { Debouncer, animateInterval } from './coreUtils/scheduler';
 export * from './data/model';
 export {
     MetadataProvider, MetadataCanConnect, MetadataCanModifyEntity, MetadataCanModifyRelation,
-    MetadataEntityTypeShape,
+    MetadataEntityShape, MetadataPropertyShape,
 } from './data/metadataProvider';
 export {
     ValidationProvider, ValidationEvent, ValidationResult, ValidatedElement, ValidatedLink,